% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{print.boot_est}
\alias{print.boot_est}
\title{print bootstrap}
\usage{
\method{print}{boot_est}(x, probs = c(0.025, 0.975), ...)
}
\arguments{
\item{x}{a boot_est object}

\item{probs}{A list of quantiles for the bootstrap}

\item{...}{additional parameters for print.data.frame}
}
\description{
print bootstrap
}
