% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{srs_bootstrap}
\alias{srs_bootstrap}
\title{Performs a bootstrap assuming a simple random sample}
\usage{
srs_bootstrap(dat, fun, replicates = 1000)
}
\arguments{
\item{dat}{a data.frame}

\item{fun}{a function taking a dataframe as a parameter}

\item{replicates}{the number of replicates}
}
\value{
a list with class boot_est containing
'value': the value of the function applied to dat.
'var': the boostrap variance.
'replicates': The bootstrap values of fun.
'nrep': the number of replicates.
}
\description{
Performs a bootstrap assuming a simple random sample
}
\examples{
df <- data.frame(a=rnorm(100))
srs_bootstrap(df, colMeans,replicates=100)$var
}

