% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{survey_bootstrap}
\alias{survey_bootstrap}
\title{Perorm a survey bootstrap}
\usage{
survey_bootstrap(design, fun, ...)
}
\arguments{
\item{design}{an object of class svydesign}

\item{fun}{a function taking a dataframe as a parameter}

\item{...}{additional parameters passed to as.srvrepdesign}
}
\value{
a list containing
'value': the value of the function applied to dat.
'var': the boostrap variance.
'replicates': The bootstrap values of fun.
'nrep': the number of replicates.
}
\description{
Perorm a survey bootstrap
}
\examples{
library(survey)
data(api,package="survey")
## one-stage cluster sample
dclus1<-svydesign(id=~dnum, weights=~pw, data=apiclus1, fpc=~fpc)
## convert to JK1 jackknife
rclus1<-as.svrepdesign(dclus1)
## convert to bootstrap
set.seed(1)
bclus1<-as.svrepdesign(dclus1,type="bootstrap", replicates=100)
attr(svymean(~api00, bclus1),"var")
set.seed(1)
survey_bootstrap(dclus1, fun=function(dat, wts) {sum(wts*dat$api00) / sum(wts)},
  type="bootstrap", replicates=100)$var
}

