% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{tstdat}
\alias{tstdat}
\title{A Synthetic Dataset for Incidence Estimation}
\usage{
data(tstdat)
}
\description{
A Synthetic Dataset for Incidence Estimation
}
\section{Variables}{

'hiv' - HIV status
'ever_test' - Reported having ever had an HIV test.
'last_test' - number of months since last HIV test.
'undiag' - Actually has never been diagnosed with HIV in the past.
'report_pos' - Reports having a previous positive test.
'biomarker_art' - Tests positive for ART biomarkers.
'low_viral' - Is classified as having low (undetectable) viral load.
}

\keyword{datasets}
