% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{bootstrap_incidence}
\alias{bootstrap_incidence}
\title{Calculates bootstrap variability estimates}
\usage{
bootstrap_incidence(incidence, nrep = 1000, rep_weights = NULL,
  type = c("BRR", "Fay", "JK1", "JKn", "bootstrap", "other"),
  combined_weights = TRUE, show_progress = TRUE, ...)
}
\arguments{
\item{incidence}{a test_inc object returned by testing_incidence.}

\item{nrep}{The number of bootstrap replications (ignored if rep_weights is not null.}

\item{rep_weights}{A dataframe of replicate weights.}

\item{type}{The type of resampling weights. See svrepdesign.}

\item{combined_weights}{TRUE if the rep_weights already include the sampling weights. This is usually the case.}

\item{show_progress}{print bootstrap progress}

\item{...}{additional parameters to svrepdesign.}
}
\description{
Calculates bootstrap variability estimates
}
