% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{bootstrap_incidence}
\alias{bootstrap_incidence}
\title{Bootstrap variance for incidence}
\usage{
bootstrap_incidence(frame, id_formula, strata_formula, weights_formula,
  report_hiv_pos, biomarker_art, low_viral, hiv, last_hiv_lower, last_hiv_upper,
  ever_hiv_test, replicates = 5000, type = "mrbbootstrap", ...)
}
\arguments{
\item{frame}{a dataframe conatining the variables needed for id_formula, strata_formula and wieghts_formula}

\item{id_formula}{a survey formula specifying the unit of sampling}

\item{strata_formula}{a survey formula specifying the strata structure}

\item{weights_formula}{a survey formula specifying the weights}

\item{report_hiv_pos}{A logical vector indicating whether each subject reported a positive hiv status}

\item{biomarker_art}{A logical vector indicating whether ART antibodies are present. NA if test not done.}

\item{low_viral}{A logical vector indicating whether viral load is <=1000}

\item{hiv}{A logical vector indicating hiv status}

\item{last_hiv_lower}{A numeric vector indicating the lower bound of the months since last hiv test bin in months.}

\item{last_hiv_upper}{A numeric vector indicating the upper bound of the months since last hiv test bin in months.}

\item{ever_hiv_test}{A logical vector indicating whether the subject had eer had an hiv test.}

\item{replicates}{The number of bootstrap replicated}

\item{type}{The type of survey bootstrap}

\item{...}{additional parameters for testing_incidence}
}
\description{
Bootstrap variance for incidence
}

