% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{testing_incidence}
\alias{testing_incidence}
\title{Incindence from testing history}
\usage{
testing_incidence(report_hiv_pos, biomarker_art, low_viral, hiv, last_hiv_test,
  ever_hiv_test, weights = rep(1,
  length(report_hiv_pos))/length(report_hiv_pos), distribution = "weibull",
  ptruth = NULL, ptreated = NULL)
}
\arguments{
\item{report_hiv_pos}{A logical vector indicating whether each subject reported a positive hiv status}

\item{biomarker_art}{A logical vector indicating whether ART antibodies are present. NA if test not done.}

\item{low_viral}{A logical vector indicating whether viral load is <=1000}

\item{hiv}{A logical vector indicating hiv status}

\item{last_hiv_test}{A numeric vector indicating the lower bound of the time since last hiv test bin in months. See details.}

\item{ever_hiv_test}{A logical vector indicating whether the subject had eer had an hiv test.}

\item{weights}{Survey weights}

\item{distribution}{Either "weibull" or "gamma." This controls the family of distribution used to
model time since last test.}

\item{ptruth}{The proportion of the diagnosed hiv positive population that would report being hiv positive.
If NULL, this is estimated using biomarker_art and low_viral.}

\item{ptreated}{The proportion of hiv positive individuals with postive biomarker_art or low_viral.
If NULL, this is estimated from the data.}
}
\description{
Incindence from testing history
}
\details{
last_hiv_test assumes that last hiv test information is captured in bins,
for example 0-6 month ago, 6-12 months ago, 12-24 month ago, and 24 or greater months ago.
Subjects in the first bin would have a last_hiv_test value of 0, those in the second would have 6,
the third would have 12 and the last would have 24. Those with missing data or who never had an hiv test
should be assigned NA.
}

